+++ src/ComplexNumber.hpp
@@ -8,14 +8,14 @@
 //===========================================================================//
 
 struct ComplexNumber {
-	double real;
-	double imag;
+	long double real;
+	long double imag;
 
 	ComplexNumber() : real(0), imag(0) {};
-	ComplexNumber(double real, double imag) : real(real), imag(imag) {};
+	ComplexNumber(long double real, long double imag) : real(real), imag(imag) {};
 
 	ComplexNumber operator+(const ComplexNumber &c);
 	ComplexNumber operator*(const ComplexNumber &c);
 
-	double abs();
+	long double abs();
 };
